% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_cause_death.R, R/try_other_cause_death.R
\name{sim_othercause_death}
\alias{sim_othercause_death}
\title{Simulate Other-Cause Death Time}
\usage{
sim_othercause_death(othercause_death_table, ID = NA)

sim_othercause_death(othercause_death_table, ID = NA)
}
\arguments{
\item{othercause_death_table}{A data frame with survival probabilities for causes other
than the selected cancers.
  Includes columns: age, sex, year, other_cause_rate, surv.}

\item{ID}{optional ID to set random seed.}
}
\value{
A numeric value representing the simulated time to death from other causes.

A numeric value representing the simulated time to death from other causes.
}
\description{
Simulates the time to death from other causes (non-MCED cancers) based on
a survival table, with optional seed setting for reproducibility.

This function simulates the time to death from causes other than the selected (MCED) cancers
using the survival probabilities from the `make_othercause_death_table` function.
}
\examples{
\dontrun{
# Simulate without setting seed
death_time <- sim_othercause_death(survival_table)

# Simulate with reproducible seed
death_time <- sim_othercause_death(survival_table, ID = 123)
}

othercause_death_table <- make_othercause_death_table(cdc_data, MCED_cdc, hmd_data, selected_cancers, "Female", 2019, 60)
sim_time <- sim_othercause_death(othercause_death_table)
}
